Move `emergencyReparenterTestTMClient` to `testutil.TabletManagerClient`
Only change was the type name, and the addition of documentation.
